[ALPS03729871] TEE: Modify Watchdata driver
[Detail]
Modify Watchdata driver from Watchdata TEE.

MTK-Commit-Id: 85be8931a502f965e45403bf4124657b6045c219

CR-Id: ALPS03729871
Feature: Watchdata TEE

+++ drivers/misc/mediatek/watchtrust/core/tee_shm.c
@@ -121,7 +121,7 @@ struct tee_shm *tee_shm_alloc(struct tee *tee, size_t size, uint32_t flags)
 	ret = sg_alloc_table_from_pages(&shm->sgt, &page,
 					nr_pages, 0, nr_pages * PAGE_SIZE,
 					GFP_KERNEL);
-	if (IS_ERR_VALUE(ret)) {
+	if (IS_ERR_VALUE((long)ret)) {
 		dev_err(_DEV(tee), "%s: sg_alloc_table_from_pages() failed\n",
 			__func__);
 		tee->ops->free(shm);
